Added new function 'getProteinInfoFromTheInternet' to check Uniprot or Entrez based on ACs
The function 'getProteinInfoFromTheInternet' checks whether the AC has
the pattern of Uniprot or Entrez Protein ACs, and submits those them
to the respective database.

Further, the ProteinGroup initializer now writes a 'database' to the
isoform_to_geneproduct table. Splice variants are only determined for
those ACs which do have the Uniprot pattern.